src/common/mempool.cc: Choose shard with sched_getcpu() to reduce cache line thashing
Improve the performance of mempool sharding on systems with sched_getcpu()

The idea is to assign threads to shards based on which CPU core they are
executing on using sched_getcpu() to select the shard. All the threads
executing on the same CPU core share the same shard. If each CPU core has its
own shard there should be no cache line contention.

Fixes: https://tracker.ceph.com/issues/64100

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>