osd: Restrict choice of primary for EC pools
Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>

Experimental changes to restrict choice of primary in
EC pools with allow_optimizations set.

Change is incomplete and includes changes that didn't
work. Only the code in _pick_primary is having an effect

The change in get_temp_osds, the new predicate and the
change in find_best_info did not manage to restrict the
primary.