osd: Add partial_writes_last_complete to pg_info_t
Data structure changes to add partial_writes_last_complete to pg_info_t.
If there is a large seqeunce of partial writes then the log may not go
back far enough to deduce that a shard has intentionally not been
receiving updates. The intent is to update state in pg_info_t as log
entries are completed.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>