osd,mon: Record nonprimary shards in pg_pool_t
EC Partial writes needs to restricy the choice of primary, a
vector is added to pg_pool_t which is pre-computed when
ec_allow_optimizations is set to restrict which shards can
be a primary. This is then used by OSDMap and PeeringState to
restrict the choice of primary and acting primary.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>