osd: Fix rollback for partial write + append
Partial write + append was generating a mod_desc rollback that
tried to do a reverse clone on more shards than the partial write.
Fix this by adding set of shards to the rollback_extents opcode.
It is possible that multiple rollback_extents need to reverse clone
different sets of shards so instead of using a vector of extents
add multiple rollback extent opcodes - one for each range. Almost
all I/Os will have 0 or 1 rollback extent ranges so no concerns that
this will bloat the log.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>