osd: ec extent cache: Use op destroy and police object deconstruct.
I was suspicous of a memory leak, but it turns out the old fix was actually good.  Problem was we were creating a loop:

1. RMW op is referenced by completion lambda
2. lambda is referenced by cache op
3. cache op is referenced by rmw op.

So even if we delete all other references, this little loop prevents any of the objects being destroyed.

However, performing the clear operation breaks the loop.

I have delivered this because:
1. I have a test for this now.
2. I prefer the ~Op mechanism.