osd: ec: Move cache into lines
Before this commit, each object held its own cache.  This cache was only discarded once all writes to that object had finished.  This is fine as long as you don't need an LRU.

With an LRU, we want to operate on finer grained chunks of data and minimise the number of operations which need to hold the lock.  Here we prepare for re-establishing the LRU by moving all caching into individual cache lines.