osd: truncate-to-grow any shards where write does not reach end.
We have introduced an optimisation which means EC does not have to zero pad holes in data buffers.

This can sometimes result in shards where the last byte written is before the assumed end of the object.  This causes issues when attempting to write those sections, as this generates clones off the end of an object.

To combat this, we use a "quirk" of the truncate op - where we can truncate an object to larger than it was before. Thes result being that Bluestore behaves as if the object is the assumed size and will invent zeros for the holes.