osd: remove read zero padding which is no longer needed.
This zero padding was added when shard objects could be smaller than assumed by EC.

This was a problem for clonerange2 ops too, so we had to find an alternative solution which grows the shard to the expected size.

This zero pad was inefficient for recovery anyway and would probably have to be improved if left in, so deleting it is a win there.