osd: Rework LRU AGAIN.
The intent is to reduce the impact of the Mutex even more. Now the LRU purely stashes the actual cache buffers, which can be restored on request.  They are stored in more efficient unordered maps. In the main code path, the only objects stashed are now the ones with inflight IO.