osd: EC: Partial Metadata Writes - present_shards
Add set of present shards to pg_log_entry_t and use this with written_shards to
distinguish between a shards that are not written to because they are absent versus
shards that are not written to because of a partial write.

If all shards that were written to and present have a copy of a log entry then a
partial write has been committed and must roll forwards. If any of these shards
are missing a copy of a log entry then it must roll backwards.

If a shard is absent and written to then recovery or backfill is required.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>