osd: Peering changes - wip
Peering changes - temporariliy disable partial write updates in info because these are recorded against a shard
and cause confusion when backfill occurs and there is a new osd with the same shard but without the log. Should be
able to get this to work by putting more checks in the peering code about a pg_shard_id's current log before allowing
it to use the partial write info.

Peering changes - fixes to proc master log to just consider the acting set rather than acting+recovery. This appears to
fix the bug, but there is still extra debug enabled here that needs to be turned off, so a bit of tidy up still
required.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>